requirements.txt
click>=8.1.8
openai>=1.0.0
boto3>=1.34.0
inquirer>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
ollama>=0.1.0
